BDK CometBFT integration now completes ABCI handshake
This changes the way our cmake system picks up gRPC and its
dependencies, to pick it up from a local installation instead
of the system.

NOTE: Still crashes the [comet] test (on gRPC server stop) but
now it gets to completing the comet Echo callbacks and finishing
the ABCI App handshake.